+++ lib/rubocop/cop/mixin/hash_node.rb
@@ -1,0 +1,14 @@
+# encoding: utf-8
+
+module RuboCop
+  module Cop
+    # Common functionality for checking hash nodes.
+    module HashNode
+      def any_pairs_on_the_same_line?(node)
+        node.children[1..-1].any? do |pair|
+          !Util.begins_its_line?(pair.loc.expression)
+        end
+      end
+    end
+  end
+end